Remove .h files, pull code into top, clean up GPIO setup.
