--- examples/adc_polled/adc.h
@@ -1,67 +1,0 @@
-/*
- * Single-File-Header for using ADC with polling
- * 03-27-2023 E. Brombaugh
- */
-
-#ifndef _ADC_H
-#define _ADC_H
-
-#include <stdint.h>
-
-/*
- * initialize adc for polling
- */
-void adc_init( void )
-{
-	// ADCCLK = 24 MHz => RCC_ADCPRE = 0: divide by 2
-	RCC->CFGR0 &= ~(0x1F<<11);
-	
-	// Enable GPIOD and ADC
-	RCC->APB2PCENR |= RCC_APB2Periph_GPIOD | RCC_APB2Periph_ADC1;
-	
-	// PD4 is analog input chl 7
-	GPIOD->CFGLR &= ~(0xf<<(4*4));	// CNF = 00: Analog, MODE = 00: Input
-	
-	// Reset the ADC to init all regs
-	RCC->APB2PRSTR |= RCC_APB2Periph_ADC1;
-	RCC->APB2PRSTR &= ~RCC_APB2Periph_ADC1;
-	
-	// Set up single conversion on chl 7
-	ADC1->RSQR1 = 0;
-	ADC1->RSQR2 = 0;
-	ADC1->RSQR3 = 7;	// 0-9 for 8 ext inputs and two internals
-	
-	// set sampling time for chl 7
-	ADC1->SAMPTR2 &= ~(ADC_SMP0<<(3*7));
-	ADC1->SAMPTR2 |= 7<<(3*7);	// 0:7 => 3/9/15/30/43/57/73/241 cycles
-		
-	// turn on ADC and set rule group to sw trig
-	ADC1->CTLR2 |= ADC_ADON | ADC_EXTSEL;
-	
-	// Reset calibration
-	ADC1->CTLR2 |= ADC_RSTCAL;
-	while(ADC1->CTLR2 & ADC_RSTCAL);
-	
-	// Calibrate
-	ADC1->CTLR2 |= ADC_CAL;
-	while(ADC1->CTLR2 & ADC_CAL);
-	
-	// should be ready for SW conversion now
-}
-
-/*
- * start conversion, wait and return result
- */
-uint16_t adc_get( void )
-{
-	// start sw conversion (auto clears)
-	ADC1->CTLR2 |= ADC_SWSTART;
-	
-	// wait for conversion complete
-	while(!(ADC1->STATR & ADC_EOC));
-	
-	// get result
-	return ADC1->RDATAR;
-}
-
-#endif
