+++ examples/i2c_oled/Makefile
@@ -1,5 +1,7 @@
 TARGET:=i2c_oled
 
+CFLAGS+=-DSTDOUT_UART
+
 include ../../ch32v003fun/ch32v003fun.mk
 
 all : flash
