+++ examples/spi_dac/Makefile
@@ -1,5 +1,7 @@
 TARGET:=spi_dac
 
+CFLAGS+=-DSTDOUT_UART
+
 include ../../ch32v003fun/ch32v003fun.mk
 
 all : flash
