+++ examples/systick_irq/Makefile
@@ -1,5 +1,7 @@
 TARGET:=systick_irq
 
+CFLAGS+=-DSTDOUT_UART
+
 include ../../ch32v003fun/ch32v003fun.mk
 
 all : flash
