+++ examples/tim1_pwm/tim1_pwm.c
@@ -9,8 +9,106 @@
 
 #include "ch32v003fun.h"
 #include <stdio.h>
-#include "t1pwm.h"
 
+/*
+ * initialize TIM1 for PWM
+ */
+void t1pwm_init( void )
+{
+	// Enable GPIOC, GPIOD and TIM1
+	RCC->APB2PCENR |= 	RCC_APB2Periph_GPIOD | RCC_APB2Periph_GPIOC |
+						RCC_APB2Periph_TIM1;
+	
+	// PD0 is T1CH1N, 10MHz Output alt func, push-pull
+	GPIOD->CFGLR &= ~(0xf<<(4*0));
+	GPIOD->CFGLR |= (GPIO_Speed_10MHz | GPIO_CNF_OUT_PP_AF)<<(4*0);
+	
+	// PC4 is T1CH4, 10MHz Output alt func, push-pull
+	GPIOC->CFGLR &= ~(0xf<<(4*4));
+	GPIOC->CFGLR |= (GPIO_Speed_10MHz | GPIO_CNF_OUT_PP_AF)<<(4*4);
+		
+	// Reset TIM1 to init all regs
+	RCC->APB2PRSTR |= RCC_APB2Periph_TIM1;
+	RCC->APB2PRSTR &= ~RCC_APB2Periph_TIM1;
+	
+	// CTLR1: default is up, events generated, edge align
+	// SMCFGR: default clk input is CK_INT
+	
+	// Prescaler 
+	TIM1->PSC = 0x0000;
+	
+	// Auto Reload - sets period
+	TIM1->ATRLR = 255;
+	
+	// Reload immediately
+	TIM1->SWEVGR |= TIM_UG;
+	
+	// Enable CH1N output, positive pol
+	TIM1->CCER |= TIM_CC1NE | TIM_CC1NP;
+	
+	// Enable CH4 output, positive pol
+	TIM1->CCER |= TIM_CC4E | TIM_CC4P;
+	
+	// CH1 Mode is output, PWM1 (CC1S = 00, OC1M = 110)
+	TIM1->CHCTLR1 |= TIM_OC1M_2 | TIM_OC1M_1;
+	
+	// CH2 Mode is output, PWM1 (CC1S = 00, OC1M = 110)
+	TIM1->CHCTLR2 |= TIM_OC4M_2 | TIM_OC4M_1;
+	
+	// Set the Capture Compare Register value to 50% initially
+	TIM1->CH1CVR = 128;
+	TIM1->CH4CVR = 128;
+	
+	// Enable TIM1 outputs
+	TIM1->BDTR |= TIM_MOE;
+	
+	// Enable TIM1
+	TIM1->CTLR1 |= TIM_CEN;
+}
+
+/*
+ * set timer channel PW
+ */
+void t1pwm_setpw(uint8_t chl, uint16_t width)
+{
+	switch(chl&3)
+	{
+		case 0: TIM1->CH1CVR = width; break;
+		case 1: TIM1->CH2CVR = width; break;
+		case 2: TIM1->CH3CVR = width; break;
+		case 3: TIM1->CH4CVR = width; break;
+	}
+}
+
+/*
+ * force output (used for testing / debug)
+ */
+void t1pwm_force(uint8_t chl, uint8_t val)
+{
+	uint16_t temp;
+	
+	chl &= 3;
+	
+	if(chl < 2)
+	{
+		temp = TIM1->CHCTLR1;
+		temp &= ~(TIM_OC1M<<(8*chl));
+		temp |= (TIM_OC1M_2 | (val?TIM_OC1M_0:0))<<(8*chl);
+		TIM1->CHCTLR1 = temp;
+	}
+	else
+	{
+		chl &= 1;
+		temp = TIM1->CHCTLR2;
+		temp &= ~(TIM_OC1M<<(8*chl));
+		temp |= (TIM_OC1M_2 | (val?TIM_OC1M_0:0))<<(8*chl);
+		TIM1->CHCTLR2 = temp;
+	}
+}
+
+/*
+ * entry
+ */
 int main()
 {
 	uint32_t count = 0;
